requirements.txt
flask==2.0.1
python-dotenv==0.19.0
pdf2image==1.16.0
pytesseract==0.3.8
python-docx==0.8.11
PyPDF2==2.10.5
spacy==3.5.3
nltk==3.8.1
scikit-learn==1.2.2
pandas==1.5.3
numpy==1.23.5
Pillow==9.5.0
pdfminer.six==20221105
python-magic==0.4.27
google-generativeai==0.3.2
SQLAlchemy==2.0.25
mysqlclient==2.2.1 